fix(diagnostics): classify earlyoom SIGTERM even when MemAvailable is healthy (#2604, #2605)
earlyoom with --prefer ^(csa|cargo|...)$ sends SIGTERM to CSA processes
even when overall MemAvailable is above the diagnostic threshold. The
previous classify_signal_kill only reported Earlyoom when
strong_memory_pressure (<5% available or cgroup OOM) was true, leaving
these prefer-list kills misclassified as unknown_signal.

Now when earlyoom is running and the exit code is 143 with a sigterm/
daemon_sigterm terminal reason, classify as Earlyoom regardless of
current MemAvailable. Updated the guidance suffix to mention the
--prefer list.

Closes #2604, closes #2605
